refactor(video-annotation): extract shared lighter-tile scene + sync hooks
